put navbarData into globalHeaderData
put navbarData into globalHeaderData
