+++ src/components/ResponsiveFooter/FooterMobile.js
@@ -11,7 +11,7 @@ const FooterStyled = styled.footer`
 
 
 const FooterContainer = styled.div`
-  max-width: 1440px;
+  max-width: 1420px;
   margin: 0 auto;
   display: flex;
   justify-content: space-between;
